% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstatscn.R
\name{checkHttpStatus}
\alias{checkHttpStatus}
\title{private function for check the http status}
\usage{
checkHttpStatus(ret)
}
\arguments{
\item{ret}{the response obj returned by httr package}
}
\value{
return nothing , but if it finds some error , it stop the script
}
\description{
private function for check the http status
}

