% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstatscn.R
\name{statscnDbs}
\alias{statscnDbs}
\title{the available dbs}
\usage{
statscnDbs()
}
\value{
a data frame with 2 columns , one is the dbcode, another is the db description
}
\description{
the available dbs in the national db
}
\examples{
 statscnDbs()
}

