% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstatscn.R
\name{statscnQueryLastN}
\alias{statscnQueryLastN}
\title{fetch the lastN data}
\usage{
statscnQueryLastN(n, lang = "zh")
}
\arguments{
\item{n}{the number of rows to be fetched}

\item{lang}{string value, one of c("zh","en")}
}
\value{
the last n rows data in the latest query
}
\description{
fetch the lastN data for the latest query, only affect the number of rows in the returned data.
This function can not be used alone , statscnQueryData() has to be called before this function
}
\examples{
\dontrun{
df=statscnQueryData('A0201',dbcode='hgnd')
df2=statscnQueryLastN(20)
}
}
