% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstatscn.R
\name{statscnQueryZb}
\alias{statscnQueryZb}
\title{the data categories}
\usage{
statscnQueryZb(zbid = "zb", dbcode = "hgnd")
}
\arguments{
\item{zbid}{the father zb/category id , the root id is 'zb'}

\item{dbcode}{which db will be queried}
}
\value{
the data frame with the sub zbs/categories , if the given zbid is not a Parent zb/category, null list is returned
}
\description{
the sub data categories for the zbid category, dbcode need to be specified, where the dbcode can be fetched by function
statscnDbs(). In the returned data frame, the column 'isParent' shows if each sub category is leap category or not
}
\examples{
 statscnQueryZb()
 statscnQueryZb('A01',dbcode="hgnd")
}

