% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstatscn.R
\name{statscnRowNamePrefix}
\alias{statscnRowNamePrefix}
\title{statscnRowNamePrefix}
\usage{
statscnRowNamePrefix(p = "nrow")
}
\arguments{
\item{p}{, how to set the rowname prefix. 
it is 'nrow' by default , and it is the only supported value currently
to unset the row name prefix, call this function with p=NULL}
}
\value{
no return
}
\description{
set the rowName prefix in the dataframe
}
\details{
in case you encounter the following error:
  Error in `row.names<-.data.frame`(`*tmp*`, value = value) : 
  duplicate 'row.names' are not allowed
you need to call this function
}
