% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstatscn.R
\name{statscnDbs}
\alias{statscnDbs}
\title{the available dbs}
\usage{
statscnDbs()
}
\value{
a data frame with 3 columns , first is the dbcode, second is the db description, third is the db description in Chinese
}
\description{
the available dbs in the national db
}
\examples{
 statscnDbs()
}
