% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstatscn.R
\name{statscnRegions}
\alias{statscnRegions}
\title{the regions in db}
\usage{
statscnRegions(dbcode = "fsnd")
}
\arguments{
\item{dbcode}{the dbcode should be some province db(fs*) , city db(cs*) or internaltional db(gj*)}
}
\value{
the data frame with all the available region codes and names in the db
}
\description{
the available regions in the specified db, it is used for query the province, city and country code generally
}
\examples{
 statscnRegions('fsnd')
 statscnRegions('csnd')
 statscnRegions('gjnd')
}

