% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstatscn.R
\name{statscnRowNamePrefix}
\alias{statscnRowNamePrefix}
\title{statscnRowNamePrefix}
\usage{
statscnRowNamePrefix(p = "nrow")
}
\arguments{
\item{p="nrow"}{, this is the only supported value, it is the default value as well}
}
\value{
no return
}
\description{
set the rowName prefix in the dataframe
in case you encounter the following error:
  Error in `row.names<-.data.frame`(`*tmp*`, value = value) : 
  duplicate 'row.names' are not allowed
you need to call this function
}

